Driving mecanics now done. Rotation does not use physics... just based off of velocity (looks decent) Need to add Renderer to handle the loading/rendering now before we finally can work on collision some more

+++ BumperCars/Car.h
@@ -43,7 +43,7 @@ private:
     double posX;
     double posY;
     int angle;
-    int rotateSpeed;
+    double rotateSpeed;
     double velocity;
     double acceleration;
